Update i-m-getting-an-error-about-a-self-signed-certificate-what-do-i-do.md
I see a lot of users with errors scanning packages, having experienced the same as a Newbie. In my case the AV program is the culprit, and since I am not behind a proxy, the standard advice is misleading and it was a long time before I got to the bottom of it. The proxy in this case is the AV program. The fix is simple though.